requirements.txt
flask
psycopg2-binary
postgres
flask_restful
flask_sqlalchemy
flask_migrate
flask_marshmallow
marshmallow-sqlalchemy
flask_wtf
